* v3.1:
  add form handlers and also typeIDs to problems
  handle empty description fields for suggestions answers and questions
  small fix to all voting when rank gets below 0

+++ gorm/suggestions.go
@@ -3,6 +3,7 @@ package gorm
 import (
 	"strconv"
 
+	"errors"
 	"github.com/golang/glog"
 	"github.com/jinzhu/gorm"
 	"work/xprincipia/backend/util"
@@ -27,12 +28,26 @@ type SuggestionForm struct {
 	Description string
 }
 
+//SuggestionDeleteForm : ~
+type SuggestionDeleteForm struct {
+	Username string
+	ID       int
+}
+
 /*
 API
 */
 
-//CreateSuggestion : Creates a question
-func CreateSuggestion(form SuggestionForm) {
+//CreateSuggestion : Creates a suggestion
+func CreateSuggestion(form SuggestionForm) error {
+
+	//Handle form Field Errors
+	switch {
+	case form.Description == "":
+		return errors.New("Description is empty: Please fill in field")
+	}
+
+	//Create Suggestion
 	s := Suggestion{}
 	s.Username = form.Username
 	intType, _ := strconv.Atoi(form.Type)
@@ -42,7 +57,8 @@ func CreateSuggestion(form SuggestionForm) {
 	s.Description = form.Description
 	s.Rank = 0
 	db.Create(&s)
-	return
+
+	return nil
 }
 
 //GetSuggestionByID : Returns a Suggestion based on an int ID
@@ -74,11 +90,26 @@ func GetAllSuggestionsByTypeID(dataType int, typeID int) []Suggestion {
 	return s
 }
 
+// UpdateSuggestion : Updates a problem with problemForm as input
+func (s *Suggestion) UpdateSuggestion(form SuggestionForm) {
+	err := db.First(&s)
+	if err == nil {
+		glog.Error("There was an error")
+	}
+
+	s.Description = form.Description
+	db.Save(&s)
+}
+
 //DeleteSuggestionByID : //DELETE
-func DeleteSuggestionByID(id int) {
+func DeleteSuggestionByID(form SuggestionDeleteForm) error {
 	s := Suggestion{}
-	s.GetSuggestionByID(uint(id))
-	db.Delete(&s)
+	s.GetSuggestionByID(uint(form.ID))
+	if s.Username == form.Username {
+		db.Delete(&s)
+		return nil
+	}
+	return errors.New("UnAuthorized User")
 }
 
 //VoteSuggestion : ~
@@ -103,21 +134,10 @@ func (s *Suggestion) VoteSuggestion(id int, vote bool) {
 
 	for i := 0; i < len(suggestions); i++ {
 		var percentRank = float32(0.0)
-		if totalVotes != 0 {
+		if totalVotes > 0 {
 			percentRank = float32(suggestions[i].Rank) / float32(totalVotes)
 		}
 		db.Model(&suggestions[i]).Update("percent_rank", percentRank)
 	}
 
-}
-
-// UpdateSuggestion : Updates a problem with problemForm as input
-func (s *Suggestion) UpdateSuggestion(form SuggestionForm) {
-	err := db.First(&s)
-	if err == nil {
-		glog.Error("There was an error")
-	}
-
-	s.Description = form.Description
-	db.Save(&s)
 }
